feat: update component analysis API to work with manifest path instead of manifest content (#165)
## Description

Updates the API for the `componentAnalysis` entrypoint to match changes made to the component analysis entrypoint in the Java API (see linked issue below). Component analysis now always operates off the manifest path (as the gradle analysis provider already did) instead of manifest content, same as stack analysis.

Note: This is a breaking change to the API.

**Related issues (if any):** https://github.com/trustification/exhort-javascript-api/issues/166

## Checklist

- [x] I have followed this repository's contributing guidelines.
- [x] I will adhere to the project's code of conduct.

## Additional information

> Anything else?