feat: support mvnw if requested and exists (#170)
## Description

Adds support for using [mvnw](https://maven.apache.org/wrapper/) instead of maven from PATH/EXHORT_MVN_PATH. If set, mvnw is searched for starting in the directory the passed manifest is in, traversing upwards until either the git repo root is hit, or the root of the drive. 

The `EXHORT_PREFER_MVNW` env var controls this preference.

**Related issues (if any):** https://github.com/trustification/exhort-javascript-api/issues/169

## Checklist

- [x] I have followed this repository's contributing guidelines.
- [x] I will adhere to the project's code of conduct.

## Additional information

> Anything else?