chore: revert to RHEcosystemAppEng for installing exhort-javascript-api in docker image (#173)
## Description

It seems even though this repo was migrated from RHEcosystemAppEng org to trustification, packages published while in the old org can't be referenced with the new orgs namespace. So we'll have to keep it for this section where we reference an old version. Speaking of, why does this reference a hard-coded old version instead of tracking main? :thinking: 

## Checklist

- [x] I have followed this repository's contributing guidelines.
- [x] I will adhere to the project's code of conduct.

## Additional information

> Anything else?